+++ lib/govuk_publishing_components/app_helpers/countdown_helper.rb
@@ -1,18 +1,18 @@
 module GovukPublishingComponents
   module AppHelpers
     class CountdownHelper
-      END_OF_TRANSITION_PERIOD = Date.new(2021, 1, 1)
+      DEADLINE = Time.new(2020, 12, 31, 23, 59)
 
       def days_left
-        sprintf "%02d", days_left_integer
+        sprintf "%02d", days_left_until_deadline
       end
 
       def show?
-        days_left_integer.positive?
+        minutes_left_until_deadline >= 30
       end
 
       def days_text
-        if days_left_integer == 1
+        if days_left_until_deadline == 1
           I18n.t!("components.transition_countdown.day_to_go")
         else
           I18n.t!("components.transition_countdown.days_to_go")
@@ -21,12 +21,28 @@ module GovukPublishingComponents
 
     private
 
-      def days_left_integer
-        (END_OF_TRANSITION_PERIOD - today_in_london).to_i
+      def days_left_until_deadline
+        (minutes_left_until_deadline / 60 / 24).ceil
       end
 
-      def today_in_london
-        Time.find_zone("Europe/London").today
+      def minutes_left_until_deadline
+        (seconds_left_until_deadline / 60)
+      end
+
+      def seconds_left_until_deadline
+        end_of_transition_period - now_in_london
+      end
+
+      def end_of_transition_period
+        london_time_zone(DEADLINE)
+      end
+
+      def now_in_london
+        london_time_zone(Time.now)
+      end
+
+      def london_time_zone(time)
+        time.in_time_zone("Europe/London")
       end
     end
   end